// src/app/models/session-raw.model.ts
export class SessionRaw {
    PatientId: number | undefined;
    SessionId: number | undefined;
    CardiacLim: number | undefined;
    PulmonaryLim: number | undefined;
    OtherLim: number | undefined;
    CardiacLimProba40: number | undefined;
    CardiacLimProba50: number | undefined;
    CardiacLimProba60: number | undefined;
    CardiacLimProba70: number | undefined;
    CardiacLimProba80: number | undefined;
    CardiacLimProba90: number | undefined;
    CardiacLimProba100: number | undefined;
    PulmonaryLimProba40: number | undefined;
    PulmonaryLimProba50: number | undefined;
    PulmonaryLimProba60: number | undefined;
    PulmonaryLimProba70: number | undefined;
    PulmonaryLimProba80: number | undefined;
    PulmonaryLimProba90: number | undefined;
    PulmonaryLimProba100: number | undefined;
    OtherLimProba40: number | undefined;
    OtherLimProba50: number | undefined;
    OtherLimProba60: number | undefined;
    OtherLimProba70: number | undefined;
    OtherLimProba80: number | undefined;
    OtherLimProba90: number | undefined;
    OtherLimProba100: number | undefined;
    constructor() {
    }
}